+++ src/Instance/InstanceMovement/InstanceMovementDetails/InstanceMovementDetailsContainer.test.js
@@ -1,6 +1,6 @@
 import '../../../../test/jest/__mock__';
 import { MemoryRouter } from 'react-router-dom';
-import { render, screen } from '@testing-library/react';
+import { render, screen } from '@folio/jest-config-stripes/testing-library/react';
 import InstanceMovementDetailsContainer from './InstanceMovementDetailsContainer';
 
 jest.mock('./InstanceMovementDetails', () => jest.fn().mockReturnValue('InstanceMovementDetails'));