+++ src/components/SearchModeNavigation/SearchModeNavigation.test.js
@@ -1,4 +1,4 @@
-import { screen } from '@testing-library/react';
+import { screen } from '@folio/jest-config-stripes/testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
 
 import '../../../test/jest/__mock__';